Move simple WHERE matching into TableMock
